requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.5
django-crispy-forms==2.0
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
psycopg==3.1.10
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
